requirements.txt
wheel==0.24.0
flask==0.10.1
flask-socketio==1.2
pyserial==2.7
pyusb==1.0.0b2
eventlet==0.17.4
pycrypto==2.6.1
